test: Added RBAC tests and more
