+++ casbin/casbin.h
@@ -24,3 +24,4 @@
 #include "persist.h"
 #include "util.h"
 #include "exception.h"
+#include "rbac.h"
