--- test/.clang-format
@@ -1,27 +1,0 @@
-# Visual Studio 生成的 .clang-format 文件
-
-BasedOnStyle: Google
-PointerAlignment: Right
-AccessModifierOffset: -4
-IndentWidth: 4
-MaxEmptyLinesToKeep: 1
-BreakBeforeBraces: Attach
-AllowShortFunctionsOnASingleLine: true
-AllowShortIfStatementsOnASingleLine: true
-AlignAfterOpenBracket: true
-IndentCaseLabels: true
-ObjCBlockIndentWidth: 4
-ObjCSpaceAfterProperty: true
-ColumnLimit: 0
-AlignTrailingComments: true
-SpaceAfterCStyleCast: false
-AlignOperands: true
-SpacesInSquareBrackets: false
-AlignConsecutiveDeclarations: false
-SpacesInContainerLiterals: false
-BreakConstructorInitializersBeforeComma: true
-AllowAllParametersOfDeclarationOnNextLine: true
-ContinuationIndentWidth: 4
-TabWidth: 4
-SpaceBeforeAssignmentOperators: true
-SpacesBeforeTrailingComments: 1
